--- a/executor/hash_table.go
+++ b/executor/hash_table.go
@@ -18,7 +18,6 @@ import (
 	"hash/fnv"
 
 	"github.com/pingcap/errors"
-
 	"github.com/pingcap/tidb/sessionctx"
 	"github.com/pingcap/tidb/sessionctx/stmtctx"
 	"github.com/pingcap/tidb/types"