<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

This PR enables merge join to spill rows with same key to disk so that the tidb server won't go OOM if there are two many rows with the same key.

### What is changed and how it works?

Created a `RowContainer` that handles rows in memory and also in disk.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test

Code changes

No user visible changes, however internally, `MergeJoinExec` and `innerJoinTable` has been changed.

Side effects

 - Possible performance regression, the regression can be minor, see benchmark below.
 - Increased code complexity.

Related changes

 - Need to cherry-pick to the release branch, should be included in release 4.0.

Release note

- We now support spilling to disk during merge join if there are some key to be joined has too many rows that surpass the capacity of the memory.