@ichn-hu 2019-12-09
/run-all-tests