@SunRunAway 2019-12-12
It is preferable to have a benchmark to show that it does not have performace regression when replacing to `rowContainer`