@ichn-hu 2019-12-26
The benchmark result:

## with master

![bench-master-flame](https://user-images.githubusercontent.com/29735669/71460224-f33bd480-27e5-11ea-956a-8b22bdca1bfd.png)

```
(pprof) top 20
Showing nodes accounting for 78.84s, 93.45% of 84.37s total
Dropped 229 nodes (cum <= 0.42s)
Showing top 20 nodes out of 54
      flat  flat%   sum%        cum   cum%
    64.42s 76.35% 76.35%     64.42s 76.35%  runtime.memmove
     3.38s  4.01% 80.36%     35.82s 42.46%  github.com/pingcap/tidb/util/chunk.(*Chunk).AppendPartialRow
     2.88s  3.41% 83.77%      2.88s  3.41%  runtime.madvise
     2.09s  2.48% 86.25%      2.09s  2.48%  runtime.memclrNoHeapPointers
     1.30s  1.54% 87.79%      1.30s  1.54%  github.com/pingcap/tidb/util/chunk.(*Column).IsNull
     1.06s  1.26% 89.05%      1.06s  1.26%  github.com/pingcap/tidb/util/chunk.(*Column).appendNullBitmap
     0.62s  0.73% 89.78%     37.49s 44.44%  github.com/pingcap/tidb/executor.(*innerJoiner).tryToMatchInners
     0.50s  0.59% 90.38%      0.50s  0.59%  runtime.futex
     0.48s  0.57% 90.94%     41.79s 49.53%  github.com/pingcap/tidb/executor.(*MergeJoinExec).joinToChunk
     0.36s  0.43% 91.37%      1.25s  1.48%  github.com/pingcap/tidb/executor.(*MergeJoinExec).compare
     0.29s  0.34% 91.72%      1.04s  1.23%  runtime.scanobject
     0.24s  0.28% 92.00%      1.41s  1.67%  github.com/pingcap/tidb/executor.(*mergeJoinInnerTable).rowsWithSameKey
     0.18s  0.21% 92.21%     36.18s 42.88%  github.com/pingcap/tidb/executor.(*baseJoiner).makeJoinRowToChunk
     0.18s  0.21% 92.43%      0.87s  1.03%  github.com/pingcap/tidb/util/chunk.Row.IsNull
     0.18s  0.21% 92.64%      2.03s  2.41%  runtime.mallocgc
     0.16s  0.19% 92.83%      0.50s  0.59%  github.com/pingcap/tidb/expression.(*Column).EvalInt
     0.15s  0.18% 93.01%      0.65s  0.77%  github.com/pingcap/tidb/executor.compareChunkRow
     0.15s  0.18% 93.18%      0.66s  0.78%  github.com/pingcap/tidb/expression.CompareInt
     0.12s  0.14% 93.33%      7.10s  8.42%  github.com/pingcap/tidb/executor.buildMockDataSource
     0.10s  0.12% 93.45%     14.80s 17.54%  github.com/pingcap/tidb/util/chunk.(*Chunk).AppendRow

```

## with rowContainer

![bench-pr-flame](https://user-images.githubusercontent.com/29735669/71460229-f6cf5b80-27e5-11ea-831e-9588e821bcb5.png)

```
(pprof) top 20
Showing nodes accounting for 75.73s, 91.27% of 82.97s total
Dropped 246 nodes (cum <= 0.41s)
Showing top 20 nodes out of 75
      flat  flat%   sum%        cum   cum%
    60.34s 72.73% 72.73%     60.34s 72.73%  runtime.memmove
     2.79s  3.36% 76.09%     32.79s 39.52%  github.com/pingcap/tidb/util/chunk.(*Chunk).AppendPartialRow
     2.71s  3.27% 79.35%      2.71s  3.27%  runtime.memclrNoHeapPointers
     2.33s  2.81% 82.16%      2.33s  2.81%  runtime.madvise
     1.23s  1.48% 83.64%      1.23s  1.48%  github.com/pingcap/tidb/util/chunk.(*Column).IsNull
     0.99s  1.19% 84.84%      0.99s  1.19%  github.com/pingcap/tidb/util/chunk.(*Column).appendNullBitmap
     0.69s  0.83% 85.67%     22.99s 27.71%  github.com/pingcap/tidb/executor.(*innerJoiner).tryToMatchInners
     0.59s  0.71% 86.38%     16.15s 19.46%  github.com/pingcap/tidb/executor.(*mergeJoinInnerTable).rowsWithSameKey
     0.56s  0.67% 87.06%      0.56s  0.67%  github.com/pingcap/tidb/util/memory.(*Tracker).Consume
     0.52s  0.63% 87.68%      0.53s  0.64%  github.com/pingcap/tidb/util/chunk.(*Chunk).NumRows
     0.44s  0.53% 88.21%      0.44s  0.53%  runtime.futex
     0.40s  0.48% 88.69%      2.98s  3.59%  runtime.mallocgc
     0.37s  0.45% 89.14%     41.69s 50.25%  github.com/pingcap/tidb/executor.(*MergeJoinExec).joinToChunk
     0.32s  0.39% 89.53%      1.15s  1.39%  runtime.scanobject
     0.31s  0.37% 89.90%      0.43s  0.52%  runtime.heapBitsSetType
     0.28s  0.34% 90.24%      0.65s  0.78%  github.com/pingcap/tidb/util/chunk.(*iterator4RowContainer).Current
     0.24s  0.29% 90.53%      0.76s  0.92%  github.com/pingcap/tidb/executor.compareChunkRow
     0.23s  0.28% 90.80%      1.44s  1.74%  github.com/pingcap/tidb/util/chunk.(*iterator4RowContainer).Next
     0.21s  0.25% 91.06%      0.54s  0.65%  github.com/pingcap/tidb/executor.(*mergeJoinInnerTable).nextRow
     0.18s  0.22% 91.27%      0.85s  1.02%  github.com/pingcap/tidb/util/chunk.Row.IsNull

```