@ichn-hu 2019-12-26
```
benchstat master.txt disk.txt
name                                                                                                                                                old time/op    new time/op    delta
MergeJoinExec/merge_join_(outerRows:300000,_innerRows:300000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8     558ms ± 5%     521ms ± 4%  -6.63%  (p=0.032 n=5+5)
MergeJoinExec/merge_join_(outerRows:3000,_innerRows:300000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8       325ms ± 2%     324ms ± 2%    ~     (p=0.841 n=5+5)
MergeJoinExec/merge_join_(outerRows:30,_innerRows:300000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8         320ms ± 2%     330ms ± 2%  +3.36%  (p=0.032 n=5+5)

name                                                                                                                                                old alloc/op   new alloc/op   delta
MergeJoinExec/merge_join_(outerRows:300000,_innerRows:300000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8    68.0MB ± 0%    68.0MB ± 0%    ~     (p=0.841 n=5+5)
MergeJoinExec/merge_join_(outerRows:3000,_innerRows:300000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8      57.9MB ± 0%    57.9MB ± 0%    ~     (p=0.952 n=5+5)
MergeJoinExec/merge_join_(outerRows:30,_innerRows:300000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8        58.6MB ± 0%    58.6MB ± 0%    ~     (p=0.683 n=5+5)

name                                                                                                                                                old allocs/op  new allocs/op  delta
MergeJoinExec/merge_join_(outerRows:300000,_innerRows:300000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8      304k ± 0%      304k ± 0%    ~     (p=0.992 n=5+5)
MergeJoinExec/merge_join_(outerRows:3000,_innerRows:300000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8       4.11k ± 0%     4.11k ± 0%    ~     (p=0.857 n=5+5)
MergeJoinExec/merge_join_(outerRows:30,_innerRows:300000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8           764 ± 0%       764 ± 0%    ~     (p=0.643 n=5+5)
```