@ichn-hu 2019-12-27
/run-all-tests