@ichn-hu 2019-12-27
benchmark result:
```
(base)  ✘ ichn@ichn-arch-pc  /tmp/benchmark  benchstat master4/master.txt disk4/disk.txt
name                                                                                                                                                old time/op    new time/op    delta
MergeJoinExec/merge_join_(outerRows:300000,_innerRows:300000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8     541ms ± 8%     613ms ± 9%       ~     (p=0.095 n=5+5)
MergeJoinExec/merge_join_(outerRows:3000,_innerRows:300000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8       331ms ± 2%     341ms ± 1%     +3.08%  (p=0.016 n=5+4)
MergeJoinExec/merge_join_(outerRows:30,_innerRows:300000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8         318ms ± 4%     322ms ± 2%       ~     (p=0.421 n=5+5)
MergeJoinExec/merge_join_(outerRows:300000,_innerRows:330000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8     586ms ±11%     605ms ± 6%       ~     (p=0.841 n=5+5)

name                                                                                                                                                old alloc/op   new alloc/op   delta
MergeJoinExec/merge_join_(outerRows:300000,_innerRows:300000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8    68.0MB ± 0%    76.9MB ± 0%    +13.12%  (p=0.008 n=5+5)
MergeJoinExec/merge_join_(outerRows:3000,_innerRows:300000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8      57.9MB ± 0%    65.9MB ± 0%    +13.72%  (p=0.008 n=5+5)
MergeJoinExec/merge_join_(outerRows:30,_innerRows:300000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8        58.6MB ± 0%    64.5MB ± 0%     +9.98%  (p=0.008 n=5+5)
MergeJoinExec/merge_join_(outerRows:300000,_innerRows:330000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8    68.0MB ± 0%    76.9MB ± 0%    +13.10%  (p=0.008 n=5+5)

name                                                                                                                                                old allocs/op  new allocs/op  delta
MergeJoinExec/merge_join_(outerRows:300000,_innerRows:300000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8      304k ± 0%      608k ± 0%    +99.94%  (p=0.016 n=4+5)
MergeJoinExec/merge_join_(outerRows:3000,_innerRows:300000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8       4.11k ± 0%    33.44k ± 0%   +713.11%  (p=0.008 n=5+5)
MergeJoinExec/merge_join_(outerRows:30,_innerRows:300000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8           764 ± 0%      8543 ± 0%  +1018.46%  (p=0.008 n=5+5)
MergeJoinExec/merge_join_(outerRows:300000,_innerRows:330000,_concurency:4,_outerJoinKeyIdx:_[0_1],_innerJoinKeyIdx:_[0_1],_NeedOuterSort:false)-8      304k ± 0%      608k ± 0%    +99.93%  (p=0.008 n=5+5)
```